Misc clean up

+++ src/main/java/game/components/holdables/weapons/Magazine.java
@@ -44,6 +44,6 @@ public class Magazine implements IMagazine {
 
   @Override
   public float getFractionFilled() {
-    return (float) ammo / (float) size;
+    return ammo / (float) size;
   }
 }
